Converted parse_host_registry into pseudo factory method,
